# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Simple CRUD",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local"
        }
    ],
    "paths": {
        "/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get access token for application",
                "description": "Return access token",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/AuthResponse"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Credential not match",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/AuthRequest"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/AuthRequest"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/auth/logout": {
            "delete": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout from application",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Handle user logout from application",
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Wrapper"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Handle user logout from application",
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthRequest": {
                "title": "Authentication Request",
                "description": "AuthResponse request model",
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email address",
                        "type": "string",
                        "format": "string",
                        "example": "user@test.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string",
                        "format": "string",
                        "example": "password"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AuthRequest"
                }
            },
            "AuthResponse": {
                "title": "Authentication",
                "description": "AuthResponse model",
                "properties": {
                    "accessToken": {
                        "title": "access_token",
                        "description": "access token",
                        "type": "string",
                        "format": "string",
                        "example": "asdasdasdsadasd"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AuthResponse"
                }
            },
...
            "Wrapper": {
                "title": "Wrapper Response",
                "description": "Wrapper Response",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Data",
                        "type": "object",
                        "format": "object"
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "format": "string",
                        "example": "message"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Wrapper"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}